More static_cast calls

+++ src/CartBlock.C
@@ -112,7 +112,8 @@ void CartBlock::getInterpolatedData(
             for (i = 0; i < listptr->nweights; i++) {
                 int const cell_index =
                     static_cast<int>(cart_utils::get_cell_index(
-                        dims[0], dims[1], nf, listptr->inode[3 * i],
+                        dims[0], dims[1], nf,
+                        listptr->inode[static_cast<int>(3 * i)],
                         listptr->inode[3 * i + 1], listptr->inode[3 * i + 2]));
                 for (n = 0; n < nvar_cell; n++) {
                     weight = listptr->weights[i];
@@ -169,7 +170,8 @@ void CartBlock::preprocess(CartGrid* cg)
     for (int n = 0; n < 3; n++) {
         dx[n] = cg->dx[3 * global_id + n];
     }
-    dims[0] = cg->ihi[3 * global_id] - cg->ilo[3 * global_id] + 1;
+    dims[0] = cg->ihi[static_cast<int>(3 * global_id)] -
+              cg->ilo[static_cast<int>(3 * global_id)] + 1;
     dims[1] = cg->ihi[3 * global_id + 1] - cg->ilo[3 * global_id + 1] + 1;
     dims[2] = cg->ihi[3 * global_id + 2] - cg->ilo[3 * global_id + 2] + 1;
     nf = cg->nf;
